// src/components/ui/Modal/types.ts
export type ModalSize = 'small' | 'medium' | 'large' | 'fullscreen'

export type ModalProps = {
  isOpen: boolean
  onClose: () => void
  children: React.ReactNode
  size?: ModalSize
  showOverlay?: boolean
  showCloseButton?: boolean
  title?: React.ReactNode
  footer?: React.ReactNode
  className?: string
}